Changed Stat attribute points from Integer to double.

+++ grails-app/domain/com/traderapist/models/Stat.groovy
@@ -7,7 +7,7 @@ class Stat {
     Integer week
     Integer statKey
     Integer statValue
-    Integer points
+    double points
 
     static transients = ['points']
